requirements.txt
flask
python-binance
pandas
pandas-ta
requests
pytz
